requirements.txt
nextcord
coloredlogs
beautifulsoup4
argparse
python-dotenv
asyncpraw
requests>2.28.1
emoji
aiohttp
